More boolean to byte conversions. Declare stuff as const where required. Optimize T_MovePlane, P_ChangeSector, P_ThingHeightClip, P_PathTraverse. More memcpy optimizations.

+++ p_spec.c
@@ -404,9 +404,7 @@ int P_FindMinSurroundingLight(sector_t *sector,
 // Called every time a thing origin is about
 //  to cross a line with a non 0 special.
 //
-void P_CrossSpecialLine(int linenum,
-						int side,
-						mobj_t *thing)
+void P_CrossSpecialLine(int linenum, byte side, mobj_t *thing)
 {
 	line_t *line;
 	int ok;
